+++ src_3_0_0/modules/controls/stateControls.js
@@ -15,10 +15,8 @@ const state = {
         button3d: Button3d,
         fullScreen: FullScreen
     },
-    mobileHiddenControls: [
-        "fullScreen"
-    ],
-    menuControls: ["button3d"]
+    mobileHiddenControls: ["fullScreen"],
+    menuControls: []
 };
 
 export default state;